fixed flask host & temp_filepath not defined error

+++ app/api.py
@@ -24,6 +24,7 @@ class Upload(Resource):
 
     def post(self):
         try:
+            temp_filepath = None
             file = request.files['image']
             if file and self.allowed_file(file.filename):
                 temp_filename = str(uuid.uuid4()) + extract_file_ext(file.filename)
@@ -39,16 +40,16 @@ class Upload(Resource):
             return jsonify(vars(Upload.fail_response))
         finally:
             try:
-                print('removing:', temp_filepath)
-                os.remove(temp_filepath)
+                if temp_filepath:
+                    print('removing:', temp_filepath)
+                    os.remove(temp_filepath)
             except Exception as e:
                 print(e)
 
-if __name__ == '__main__':
-    host = 'localhost'
-    port = 8000
-    app = Flask(__name__)
+host = '0.0.0.0'
+port = 8000
+app = Flask(__name__)
 
-    api = Api(app)
-    api.add_resource(Upload, '/api/upload')
-    app.run(host=host, port=port)
+api = Api(app)
+api.add_resource(Upload, '/api/upload')
+app.run(host=host, port=port)
